fix performance issue when parsing queries
Based on commit 0bd7b8eac16e6966dea81b8a5991e052192a6af2 by Vlad Mihalcea which fixes this issue (HHH-4959) on Hibernate 5 and above.
Difference is that I used a static field with a setter to toggle the new behaviour on or off instead of SessionFactoryOptions. Mostly because I didn't want to waste time backporting changes in that area.